requirements.txt
fastapi
uvicorn
pydantic
requests
fpdf
tavily-python
